requirements.txt
asgiref==3.5.0
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.6
djangorestframework==3.12.4
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
